Phase 1 of move
