+++ projects/hipdnn/plugin_sdk/include/hipdnn_plugin_sdk/EnginePluginApi.h
@@ -6,7 +6,7 @@
 #include <stdint.h>
 
 #include <hip/hip_runtime.h>
-#include <hipdnn_sdk/plugin/PluginApi.h>
+#include <hipdnn_plugin_sdk/PluginApi.h>
 
 /**
  * @file EnginePluginApi.h
