+++ projects/hipdnn/plugin_sdk/include/hipdnn_plugin_sdk/PluginLastErrorManager.hpp
@@ -5,9 +5,9 @@
 
 #include <string>
 
+#include <hipdnn_plugin_sdk/PluginApi.h>
+#include <hipdnn_plugin_sdk/PluginDataTypeHelpers.hpp>
 #include <hipdnn_sdk/logging/Logger.hpp>
-#include <hipdnn_sdk/plugin/PluginApi.h>
-#include <hipdnn_sdk/plugin/PluginDataTypeHelpers.hpp>
 #include <hipdnn_sdk/utilities/StringUtil.hpp>
 
 // NOTE: The last_error variable must be defined in one of the plugin source files.
@@ -15,7 +15,7 @@
 // NOLINTNEXTLINE
 // thread_local char PluginLastErrorManager::s_lastError[HIPDNN_PLUGIN_ERROR_STRING_MAX_LENGTH] = "";
 
-namespace hipdnn_plugin
+namespace hipdnn_plugin_sdk
 {
 
 class PluginLastErrorManager
@@ -54,4 +54,4 @@ public:
     }
 };
 
-}
+} // namespace hipdnn_plugin_sdk
