--- projects/hipdnn/plugin_sdk/include/hipdnn_plugin_sdk/PluginSdk.hpp
@@ -1,12 +1,0 @@
-// Copyright © Advanced Micro Devices, Inc., or its affiliates.
-// SPDX-License-Identifier:  MIT
-
-#pragma once
-
-namespace hipdnn::plugin_sdk
-{
-inline void hello()
-{
-    // Placeholder function
-}
-}
